requirement.txt
numpy==1.22.4
Django==4.0.4
django-phonenumber-field[phonenumbers]==6.1.0
twilio==7.9.2 